Create models to represent most basic concepts
- Data models are free of business logic

- The Game model represents either a regular game or a tiebreak
game. It keeps point tallies and game attributes, such as deuce or
advantage. Deuce and advantage will remain unused for tiebreak. For a tiebreak,
it can be seen as deuce is always false and there is no advantage
player.

- The MatchSet model maintains a list of games and the set scores between
two sides. It also maintains if the game being played is regular or
tiebreak.