Add the Match entity
- The entities are a higher level object than the services in that
they are responsible for maintaining states and they usually invoke
services to manipulate data models.

- The entities should be well encapsulated and they can only be
interacted with by their well defined interfaces.

- The Match entity represents a single set tennis match with two
players and it records the winner of each point,
`pointWonBy`. Behind the scene, it maps the players to the internal
GameSide model and invokes the right services to fulfil business
logic.